embeddings/requirements.txt
fastapi==0.95
tensorflow_hub==0.13
uvicorn==0.22
python-dotenv==1.0
tensorflow==2.12